Local_PoLL_MQM/requirements.txt
# Core runtime
aiohttp>=3.9.0
httpx>=0.27.0
tqdm>=4.60.0
python-dotenv>=1.0.0

# Analysis / visualization
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Optional but recommended objective metrics
sacrebleu>=2.4.0
unbabel-comet>=2.2.0

# Watermark / image utilities
opencv-python>=4.9.0
imwatermark>=0.0.2

# Test / development utilities
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0

# Jupyter event loop compatibility (optional)
nest-asyncio>=1.5.0

# COMET dependencies may pull torch/pytorch-lightning automatically.
